Fix GraphML output when a formula argument is given.
GraphML requires that all nodes in the graph, including leaf nodes, are defined by a <node> tag. The 'brew deps --installed' or 'brew deps --all' commands already return a separate line for each leaf formula. When a single formula is given, the output only contains a list of  dependencies. In this case, we need to add a line for each leaf node ourselves.