Only visualize dependencies one level down, instead of recursing, when passing the --installed and --all switches.
Previously, also transitive dependencies were visualized as direct dependencies. For example, "python" is shown as a direct dependency of "graphviz" even though it is actually a dependency of "glib" which is a dependency of "gts" which is a dependency of "graphviz". This made the resulting graphs fairly confusing.

Note: The graphs of specific formulae are still visualized as before because there's no easy way to get the tree of dependencies for a specific formula. One could get the dependencies recursively by executing "brew deps <formula>" multiple times, but that might result in lots of calls. There's also "brew deps --tree" but the output would need to be parsed.